Missed two int/unsigned usages in BSSourceView.m

+++ Source/BSSourceView.m
@@ -158,14 +158,14 @@
     return;
   
   // go through the document until we find the NSRange for the line we want
-  int rangeIndex = 0;
+  NSUInteger rangeIndex = 0;
   for (NSUInteger i = 0; i < line; i++)
   {
     rangeIndex = NSMaxRange([[textView string] lineRangeForRange:NSMakeRange(rangeIndex, 0)]);
   }
   
   // now get the true start/end markers for it
-  unsigned lineStart, lineEnd;
+  NSUInteger lineStart, lineEnd;
   [[textView string] getLineStart:&lineStart end:NULL contentsEnd:&lineEnd forRange:NSMakeRange(rangeIndex - 1, 0)];
   [textView scrollRangeToVisible:[[textView string] lineRangeForRange:NSMakeRange(lineStart, lineEnd - lineStart)]];
 }
